refactor : 지역 인증 도메인 분리
