+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/areaAuth/repository/AreaAuthRepository.java
@@ -1,6 +1,6 @@
-package org.j1p5.domain.auth.area.repository;
+package org.j1p5.domain.areaAuth.repository;
 
-import org.j1p5.domain.auth.area.entity.AreaAuthEntity;
+import org.j1p5.domain.areaAuth.entity.AreaAuthEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AreaAuthRepository extends JpaRepository<AreaAuthEntity, Long> {
